fix cache stringify issue, res.status not defined
- fix fetching from db issue by first converting to .json and replacing bigints by strings, then parse back, hand over to controller which sends with .json
Previously, it already returned a JSON string, but res.json(...) in the controller expects a JS object, not a string. When this stringified data gets passed to res.json(),
the frontend receives a string instead of an array/object, which causes parsing issues or rendering errors.

- fix res.status not defined at errorMiddleware -> Express error-handling middleware must have 4 parameters,
even if you're not using next. Otherwise, Express will not treat it as an error handler, and res might be undefined or misinterpreted.
fix by adding next parameter and marking it as unused ignore for eslint